MB-47513: Don't pass loggerName to magma BucketLoggers
Currently we pass the loggerName (including the bucket name) to the
magma BucketLogger for both the name and the prefix fields. This means
that any message logged via one of the engine macros will prefix the
bucket name twice, once via the bucket name inclusion in
BucketLogger::logInner and once via the prefix. For log messages
logged via magma we use the passed loggerName as a prefix (including
the bucket name only once).

Change-Id: I13b4772cff4a4983b1f723da2180c6c311057890
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/157852
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Richard de Mellow <richard.demellow@couchbase.com>