MB-36377: Make KVStore::begin() return TransactionContext
As each type of KVStore will need a different type of TransactionContext
(queued writes will be of different types) we should return a
TransactionContext from KVStore::begin() to create the correctly typed
context for us. For now each KVStore will return an TransactionContext
but this will change as we add support to each one.

Change-Id: I821a57d2c9d9476895cc02872499f9d7436171fb
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/158214
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>