MB-47516: Remove code that renumbers all queued set-vbstate items
Code in setOpenCheckpointID is renumbering all set-vbstate items to
equal the low-seqno of the checkpoint, this break the monotonicity of
the checkpoint items. With expel running after this event, it's
possible that the checkpoint is then left in a state which returns
incorrect cursors.

Note: 7.0 (cheshire-cat) already removes this code.

Change-Id: I5f1ce82d926a9b601b9430dafecb17ec47612da9
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/158082
Reviewed-by: Dave Rigby <daver@couchbase.com>
Well-Formed: Restriction Checker
Tested-by: Build Bot <build@couchbase.com>