MB-36377: Move TransactionContext to a new file
When we move the pending reqs queues to TransactionContexts we will
do so by subclassing TransactionContext. Currently as we use
subclasses of TransactionContext (EPTransactionContext) to invoke
the persistence callback we have a problem with the class hierarchy
as logically these new TransactionContext subclasses shouldn't
subclass EPTransactionContext but then we'd need one
EPTransactionContext per TransactionContext subclass (with the
pending requests queues). To solve this problem, I'm going to include
a PersistenceCallback object in every TransactionContext, and have
subclasses of the PersistenceCallback (EPPersistenceCallback) instead
determine if we invoke the formerly name PersistenceCallback. This
requires moving TransactionContext to a new file as it will need
to include persistence_callback.h which needs to include kvstore.h.

Change-Id: I2113b6a4a04be3bfddcc164203a3a9c012bc5e00
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/158166
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>