MB-36377: Change subclassing of TransactionContext/PersistenceCallback
To allow us to create different subclasses of TransactionContext with
the pending request queues of each KVStore we need to move the current
subclassing of TransactionContext (which determines if we invoke the
PersistenceCallback or not) to a subclassing on PersistenceCallback
instead. Each TransactionContext will now have a PersistenceCallback
which may or may not be an EPPersistenceCallback which performs the
previous logic of PersistenceCallback.

Change-Id: I727f24f365efb04f19f3b2ade98907add6db1cc1
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/158213
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>