MB-47604: Separate public API of KVStore
KVStore currently serves two purposes, it acts as an API for users
of some KVStore and it included code common to all KVStore types.
To introduce an ABKVStore which runs two KVStores in parallel and
compares results of interesting functions we need to separate the
API from the common implementation so that the ABKVStore only needs
to subclass KVStoreIFace.

Change-Id: I5bfcf9e716c1b2e7f93afb7a825205798a1ccd15
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/158441
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>