MB-47197: Set user limits as part of database refresh [4/n]
Unfortunately ns_server isn't able to notify us in a proper way
when a users limit changed, but instead we're being told to
read the entire file and figure it out ourself (it could very
well just be that a new user was added to the system; a users
password change or a user being deleted).

To avoid any race conditions where users connect while we're
performing the update the new database gets installed, and then
we iterate over all of the users in the new database and tries
to update the user.

Any user connecting while we're updating the database will see
the new entry (if they haven't been inserted in the cache already)
and we'll make sure we update all exiting entries in the database.

Change-Id: I24385c32f8efd0925c9a5624a591f18ef19685e0
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/158430
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>