MB-47604: Add initial NexusKVStore implentation
For now it's just acting as a wrapper around a CouchKVStore so that
we can get the bulk of the interface committed. In a future patch
we will add the config required to choose a and b KVStore types,
create the b KVStore, and start adding comparisons of results in
interesting places.

Change-Id: Iea2160fff50a57dc10d188ebf85d9f318fe07740
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/158442
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>