MB-47197: We can't verify curr_conns right after disconnect
From the clients side it close the connection and reconnects,
but on the server side the disconnect is more than just closing
the socket. A new connection may be established on a different
front end thread, and we have no control of the scheduling
of the front end worker threads so the new connection could
perform a lot of operations while the other is currently
in the progress of shutting down.

Due to this race we cannot safely check for the number of
current connections as we don't know how far we're come
in the shutdown logic.

To verify that the counter is correct we'll check the number
of connections in the end of the test. At this time the current
number should be at max, then we close all of the "extra" connections
and run a loop up to 30 seconds to verify that the stat drops
down to 1. (30 seconds is a LONG time so we shouldn't get
any timeouts just because the server is busy)

Change-Id: Ie380ebcbb9cfd1f892f04751f37a49e332838279
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/158468
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>