MB-47604: Remove RocksDBKVStore::prepareForDeduplication
The method is empty and unused currently (as we should only be
hitting the KVStore variant - if we want to prepare for dedupe).
In a future patch when I separate the KVStore API from KVStore this
function needs to become virtual and this method currently hides the
KVStore implementation which breaks a test.

Change-Id: Id8c759fa1d5828c8bc6b5d161a890c16e1b5d776
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/158477
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>