MB-47197: Add support for MutablePasswordDatabase
The password database usage in the server is immutable. A
database is loaded of disk in one thread, and then atomically
replaced with the old version.

When running unit tests we want to be able to change the
users entries, and its easier to operate on the objects
rather than parse the json and search within that.

Change-Id: Ic1ae10435429641debe8fed33da793c2a146e223
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/158479
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>