MB-47604: Create secondary KVStore in NexusKVStore
Create the secondary KVstore in the NexusKVStore (defaults to a magma
instance). This highlights an issue where we run tests against magma
if it exists in the NexusKVStore but not stand-alone. Magma can't handle
being reconfigured with a different number of vBuckets (which is/was
being done in an ItemPager test) so we need to skip this test if we are
running the NexusKVStore with magma as we'd fail in the ctor. We should
also skip all of the ItemPager tests that are memory sensitive as
the presence of magma will cause them to fail.

Change-Id: Ia00ce0db9783febdbf96f8c73dcb3d84ae2df824
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/158594
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>