MB-47092: Move kv_datatype_count to high cardinality endpoint
kv_datatype_count represents 16 individual time series.

This metric is "sufficiently" useful that it is desirable to record it,
but not so useful as to require frequent scraping and long retention.

As explored in MB-45843, kv_datatype_count has been seen to constitute
around 3.4% (19MB/568MB) of the Prometheus disk usage in a particular
test scenario.

While this may seem small, this was the second largest metric family
present. By moving this to the high cardinality endpoint, in the same
test scenario, the disk usage of this metric would be expected to 1/6th
of the above value, or lower.

This is because the scrape interval of the HC endpoint defaults to 60s
(vs 10s for LC), and may be dynamically increased to meet a desired
retention period within a limited disk quota.

This change will have minimal impact in an idle node, as the cost
largely comes from recording multiple frequently changing values.

Change-Id: If0f2306f8bde8f16e162e0c806363f1c9a347985
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/158648
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>