MB-47604: Unify single threaded test backend config
Currently we supply "persistent" as the config for couchstore buckets
and manually handle the magma (and rocks) cases. Before adding another
type of backend, refactor the config to always include the backend in
the config string (e.g. "persistent_couchstore", "persistent_magma") so
that we can read the backend directly from the config.

Change-Id: I67d4b03271d6a26c9c823c0eb0f8861362fd326e
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/158649
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>