MB-47707: Shut down connections when the parent port dies
Disconnect all connections which is bound to a server port when the
port gets deleted. The PRD say that localhost connections should
not be affected by this so we're using a hidden setting to toggle
that mode on/off to make a test which disconnects a localhost
interface being deleted.

Change-Id: I07cbd58100ebca4b80e3ce94f0306a5825e01b11
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/158907
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>