MB-47851: Cancel any requests blocked on warmup if warmup stopped.
If warmup is cancelled before it gets past the "populateVBucketMap"
phase, and there are outstanding client requests which are blocked
until populateVBucketMap completes, then KV-Engine never notifies the
waiting client connections.

As such, any subsequent deletion of the affected bucket is blocked
forever, as the KV-Engine frontend is waiting in vain for those
connections to finish their current operation.

Solution is to ensure that we stopWarmup _before_ the bucket is
deleted (during EPEngine::initiate_shutdown()), and when warmup is
stopped we cancel any pending requests.

Change-Id: I8e94e36baf34e86fa7203973be8f73e8898be7ba
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/159901
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>
Reviewed-by: Richard de Mellow <richard.demellow@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>