MB-47811: Magma - Flush collection stats doc after meta update
As part of the collections meta update we may delete the "alive"
stats doc if we drop a collection in that flush batch. If we drop
and recreate a collection in the same flush batch though then we
are currently deleting the "alive" doc without recreating it for
the new version of the collection.

Re-order the flushing code to update the metadata (and potentially
delete the stats doc) before we flush the new version to ensure that
it makes it to disk.

Change-Id: I4ca37413ded711b5f2be95b3ae4b009d7b72de9b
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/158961
Reviewed-by: Dave Rigby <daver@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>