Remove core-specific cookie management from CookieIface
The engine should _not_ be allowed to mock around with ref counts,
blocking states etc.

In the server API we downcast to the appropriate type anyway so
we don't really need this to be a virtual interface. It is provided
as two different implementations the entire stack (server api and
the cookie ref counting) so there isn't a value from a testing
perspective.

Now that we've removed the binding we're free to flip one of them
to use smart pointers without having to deal with the other.

Change-Id: I63a1d2c1fb3834ee254d4b4c8a19d23e2cbf3076
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/159077
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>