Refactor: Don't return tenant shared_ptr from CookieIface
When a shared_ptr gets returned the actual release of the memory
isn't known (happens when the last reference goes away). By returning
a shared_ptr one may think it is safe to cache the handle, but it
could result in memory being released from the engine context
and not the core.

By returning a raw pointer (with a note that you can't cache it)
it should (hopefully) stop this from happening.

Change-Id: I209f312e61cb7439e52031388b8ae79392fa3633
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/159079
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Trond Norbye <trond.norbye@couchbase.com>