MB-47604: Add StorageProperties::PrepareCounting
To assert that the vbstate on disk is correct after calling
snapshotVBucket we want to read back the vbstate and compare the
primary against the secondary KVStore in NexusKVStore. In the case
of couchstore vs magma though, couchstore will populate the on disk
prepare and prepare bytes fields but magma currently cannot.

Add a storage property such that we can skip that comparison.

Change-Id: I8134d8cc320b2a33eba0093fc2494abe4d80bdd7
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/159142
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>