MB-47386: Move the ckpt extraction code into a dedicated function
That's the code for extracting the  closed/unref checkpoints from the
checkpoint-list, which is currently O(N) in the size of the list and
what we aim to optimize. Note that the function just returns the
extracted checkpoints to the caller, there is no deallocation.

This patch moves the code into its own function and runs a benchmark on
it. That will be our baseline for comparing the optimizations that will
come with the follow-up patch.

The benchmark results here show the current O(N) behaviour of the
function, with the runtime increasing linearly with the number of
checkpoints extracted:

$ ./ep_engine_benchmarks --benchmark_filter="ExtractClosedUnrefCheckpoints" --benchmark_repetitions=10
Run on (24 X 2400 MHz CPU s)
CPU Caches:
  L1 Data 32 KiB (x12)
  L1 Instruction 32 KiB (x12)
  L2 Unified 256 KiB (x12)
  L3 Unified 15360 KiB (x2)
-------------------------------------------------------------------------------------------------------------------
Benchmark                                                                         Time             CPU   Iterations
-------------------------------------------------------------------------------------------------------------------
CheckpointBench/ExtractClosedUnrefCheckpoints/1/iterations:1_median            5908 ns         4632 ns           10
CheckpointBench/ExtractClosedUnrefCheckpoints/10/iterations:1_median           7369 ns         5637 ns           10
CheckpointBench/ExtractClosedUnrefCheckpoints/100/iterations:1_median          8408 ns         6721 ns           10
CheckpointBench/ExtractClosedUnrefCheckpoints/1000/iterations:1_median        40237 ns        37663 ns           10
CheckpointBench/ExtractClosedUnrefCheckpoints/10000/iterations:1_median      734348 ns       728692 ns           10

Change-Id: Id2bb85bc71f3c259d075682a3dbae406ed286539
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/159194
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Paolo Cocchi <paolo.cocchi@couchbase.com>