MB-47604: Add NexusKVStore getMulti (bgFetch) support
Add support for bg fetching to NexusKVStore. One intersting thing
going on here is that KVStores may not necessarily return compressed
items if we ask for them if the value they originally stored was
uncompressed. As such, we can't make direct Item comparisons as the
datatype and value fields will differ and may need to decompress the
values to compare the results.

Change-Id: Id97d7e25475cf30b0240c64f9a765927d6a70355
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/159249
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>