MB-47790: Propagate KVStore::scan errors from DCPBackfillDisk
If a KVStore detects corruption during a scan, it fails and an error
is returned to the caller. However in the case of DCPBackfillDisk, it
did not check for such an error and instead incorrectly assumed the
scan had completed successfully. This resulted in a rebalance hang as
no more mutations were sent by the Producer to complete the backfill.

Address this by correctly propogating the error back to the DCP stream
by setting it to dead(BACKFILL_FAIL).

Change-Id: I8ce618bceaf8e140adb0a94f6c746440b913202f
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/159767
Reviewed-by: Jim Walker <jim@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>