MB-47009: When resuming a filtered snapshot, check we reach the end
In the MB a stream is resumed, but has a snapshot end greater than the
start-seqno, if the mutations in the range start to snap-end are
filtered away the stream remains dormant and the client assumes the
end has yet to be reached.

We now check for this situation when generating a memory snapshot and
ensures snapshot+seqno advance gets the client to the snap-end seqno.

MB-47534 is also addressed by this change as lastReadSeqno is updated
by any seqno-advanced sent, we don't keep sending them. An updated
unit test covers the issue seen in MB-47534.

MB-47753 is also addressed by this change. With lastReadSeqno updated we
can avoid generating extra backfills that produce empty snapshots for
already sent seqnos.

Change-Id: I152c193d805787ceeb7d43820404336bdef93604
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/159018
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>