MB-48128: Fail MagmaKVStore getCollectionsManifest if NotExists
NotExists means that the KVStore does not exist (rather than
NotFound which means that the doc was not found). Currently a
NotExists error results in an empty manifest being returned with
the bool result set to true. This change brings the functions
behaviour in line with the corresponding function in CouchKVStore.

Change-Id: I9ba86c8a99e1bce9353b05919e4d7dbbee3d6802
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/159532
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>