MB-47437: 1/n Update KVStore::GetCollectionStats for NotFound
Update the API so the caller can detect

Success - stats for the collection were found
NotFound - stats for the collection were not found
Failed - something went wrong

The NotFound case will be useful for MB-47437, but no callers
in this patch set are adjusted for special handling of the
NotFound case. Callers are already handling the case where
NotFound and stats are returned as 0.

Change-Id: I8ddbf8604c54ad053a5a298ee08e44236396c299
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/159680
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>