MB-47991: Make Magma SyncDelete meta behave like couchstore
Currently a SyncDelete prepare generates metadata in which the
deleted bit is set. This differes from couchstore in which an
additional prepare only metadata bit is set to distinguish
SyncDelete prepares from SyncWrite prepares. This results in 3
issues in MagmaKVStore:

1) The FlushState returned is Insert (as the old item is deleted)
   for MagmaKVStore whilst it is Update for couchstore if we modify
   a SyncDelete prepare. This only affects stats.

2) Magma reads the expiry field of items and if we set the item to
   deleted adds the metadata purge interval to this when considering
   it it should compact some SST to force expiry or tombstone purging.
   This currently returns in-accurate values.

3) MagmaKVStore currently purges SyncDelete prepares when they are
   older than the metadata purge interval. This means that they will
   generally be kept for far longer, but may also be purged if
   incomplete if the cluster remains unhealthy for the duration of the
   metadata purge interval. This breaks durability guarantees.

Whilst we could correct this by changing the way that we deal with
SyncDelete prepares in MagmaKVStore it means that we would behave
differently to CouchKVStore. As much of MagmaKVStore is a copy of
CouchKVStorem, it feels simpler to simply make them behave in the same
way.

MB-47958: Magma - Don't attempt to expire prepares

We should never attempt to expire prepares (and we don't in
CouchKVStore). This code was never actually being executed before
as SyncDelete prepares were already considered deleted and we
considered them for purging rather than expiration in the compaction
callback. This is already covered by unit tests by the test
DurabilityEPBucketTest.DoNotExpirePendingItem that fails without this
change.

Change-Id: I435ac8fdc9f31dec02dae5ad408ebc9a12930817
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/159531
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>