MB-51530: BP[MB47904] Always look up SID/CID when EUID is used
As an optimization to work around the regression tracked in
MB-39594 we would bypass looking up SID and CID if the privilege
context had full access to the entire bucket.

When Impersonate is being used we don't have the privilege context
for the effective user (which may _NOT_ have full access to
to the bucket) so we need to look up SID and CID so that it is
available later on when we do have that privilege context and
perform the privilege checks.

Change-Id: I0b028f21bf92d115a22b4a0bb25b695a94544b9f
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/173399
Well-Formed: Restriction Checker
Reviewed-by: Daniel Owen <owend@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>