MB-47437: 5/n Set the snapshot end correctly for a legacy stream
WARNING this fix is for persistent couchstore buckets only, the bug
is still present for ephemeral buckets.

Primary purpose of this fix to call ActiveStream::markDiskSnapshot
with the correct end/maxVisibleSeqno so that a legacy DCP stream against
a couchstore bucket isn't given a range that ends on.

* an item of another collection
* an !committed default collection item

The newly added DCPBackfillBySeqnoDisk::markLegacyDiskSnapshot does
the bulk of the work for this commit and the function comment describes
how it works and its limitations.

Change-Id: I7d27424856706b6d09e6bf4c390ce15cbd1fb991
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/159693
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>