MB-47604: Add config for expiration from start of compaction
We can't assert that the expiries generated by each KVStore in
NexusKVStore are the same if the timepoint from which we can
expire changes as we visit each item. Add configuration such that
NexusKVStore can have items be expired from the start time of each
compaction rather than a moving time point.

Change-Id: Ida6e3382c2c30063582d2bae008a6440dfa72663
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/160231
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>