MB-48038: Fix EPStats::cursorMemoryFreed
That stat is currenlty storing a quantity that has no actual meaning.

The general idea is that cursor-dropping itself doens't free any memory.
It just makes checkpoints eligible for removal, but those checkpoints
may not be released at all by the CheckpointRemover (eg, after dropping
the cursor a new cursor is registered that makes a checkpoint referenced
again).

So, the stats is renamed to EPStats::memFreedByCheckpointRemoval and
it's updated when checkpoints are actually released.

Change-Id: I1061081e28cd89ec2714cc059493db59392011f4
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/159759
Tested-by: Paolo Cocchi <paolo.cocchi@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>