MB-48034: Pass vbid to magma compaction context
Implcit magma compactions currently construct a MagmaCompactionCB
which is passed to every item visited. When we visit the first item
during an implicit compaction we create a CompactionContext object
(which typically holds the vbid that we care about) by reading the
vbid from the metadata Slice. This means that for every item we are
currently storing a vbid (16 bits) just to work out which vBucket we
are compacting. To remove the vbid from the metadata we can:

1) Have magma pass the KVStoreID(Vbid) into the function that creates
the MagmaCompactionCB so that we can store it in the MagmaCompactionCB
2) Use the vbid of the MagmaCompactionCB rather than the vbid of the
metadata Slice in MagmaKVStore::compactionCallback

In a future change I will remove the vbid from the magma metadata.

Change-Id: I0593bb6142666771bc6c7a024c0576c060b657d1
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/159860
Reviewed-by: Richard de Mellow <richard.demellow@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>