MB-48120: Make magma metadata usable in non-kv projects
Currently magma MetaData includes Item and a couple of other
kv types that we won't want to include in non-kv projects(magma)
when we re-use the metadata.

In addition, move the metadata files to a new directory that,
for want of a better name, is called "shared" so that in a future
patch we can compile the metadata in a separate library that magma
can link against.

Magma also wants the Metadata type to be convertible to and from
nlohmann json so add functions to do so.

Change-Id: I89a348f2a96a0e97de81f0915c74a5ad6da80cb9
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/159978
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>