MB-47952,MB-48121: Fix expiry of temp items
In the original change we prevented temp items from always being
expired by the item pager. If the temp items were still expired
by time though then the expiry pager would still attempt to
"expire" them. The Nexus variant of the test was hitting this
fairly frequently as it runs slower than the couchstore and magma
variants.

Correct the test to time travel so the item is always expired
(by time) and the code to never attempt to expire temp items in the
expiry pager.

Change-Id: I8b7f96bbcd0795d5688ee8731efb0752832ed48a
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/160109
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>