Refactor getConnection
Modify getConnection so that it is no longer virtual, but
the prepare call (which sets the various features) is virtual.

This is a preparation for starting to migrate over to a new
getConnection which allows for authentication and bucket selection
which will make the code simpler now that we want to get rid
of the "default bucket" logic.

Change-Id: Id24602f39eff14cbb0808f7ebbd4e13973659666
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/160119
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>