Re-apply: MB-48038: Don't trigger any CheckpointVisitor if mem-recovery not needed
Note: This patch was originally reverted as it caused MB-48233. This
re-push contains the fix for MB-48233 too, covered by test.

Original commit message:

That was our original intention in the CheckpointRemoverTask, but
actually a CheckpointVisitor run is currently triggered regardless of
any memory state.
Note that checkpoint mem-recovery triggers are now defined by the
checkpoint quota and other marks in configuration, so that's what we
want to enforce.

This patch has already good test coverage. As soon as I start enforcing
the checkpoint mem-recovery triggers a bunch of ep_testsuite tests
start failing as they assume that eventually items will be removed from
checkpoints, regardless of any memory condition. The "fix" for those
tests is also an example on how we can switch from a relaxed
mem-recovery behaviour to an eager one by just tuning the existing
configuration parameters.

Change-Id: I5fbab691e3ab8fe6fa4ec9ef38032d72d85d60e2
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/160885
Reviewed-by: Jim Walker <jim@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>