MB-48038: Don't trigger any CheckpointVisitor if mem-recovery not needed
That was our original intention in the CheckpointRemoverTask, but
actually a CheckpointVisitor run is currently triggered regardless of
any memory state.
Note that checkpoint mem-recovery triggers are now defined by the
checkpoint quota and other marks in configuration, so that's what we
want to enforce.

This patch has already good test coverage. As soon as I start enforcing
the checkpoint mem-recovery triggers a bunch of ep_testsuite tests
start failing as they assume that eventually items will be removed from
checkpoints, regardless of any memory condition. The "fix" for those
tests is also an example on how we can switch from a relaxed
mem-recovery behaviour to an eager one by just tuning the existing
configuration parameters.

Change-Id: I91cf08c0d2d061394e40e4310f90b791f2e510a7
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/160153
Tested-by: Paolo Cocchi <paolo.cocchi@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>