MB-48172: Don't skip system events in magma rollback callback
Before calling the RollbackCB we process the items being rolled back
by magma in MagmaKVStore. Currently there we are skipping SystemEvents.
This has no functional affect as these items are skipping in the
EPDiskRollbackCB but we shouldn't be asusming that the RollbackCB
doesn't want to see them, and we need to pass them to the RollbackCB
to make assertions about the values rolled back in NexusKVStore.

Change-Id: I6db279814aa966d6984309a8dbded6fbba9c4819
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/160235
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>