MB-48038: Add ability to stop execution in CheckpointVisitor
Here and in the following patches we'll be moving all the checkpoint
memory-release logic from the CheckpointRemoverTask to the
VBCBAdaptor/CheckpointVisitor fixture.

While the short-term goal is to get rid of some complexity in the
CheckpointRemoverTask, our medium-term goal also to remove any
periodic scheduling of the task for avoiding the current "busy
polling" model that keeps CPUs busy for nothing. Note that the
CheckpointRemoverTask is just one among many tasks that follow the same
model.

So in the short-term:
- I move most of the logic that actually releases checkpoint-memory from
  the CheckpointRemoverTask to the CheckpointVisitor
- Between CheckpointRemoverTask and VBCBAdaptor/CheckpointVisitor we
  keep the latter beacuse that already implements the
  PausableVBucketVisitor execution-yield mechanism that we want to
  maintain
- For now we cannot remove completely the CheckpointRemoverTask as we
  need the periodic scheduling until we implement the proper
  (event-driven) trigger logic for the CheckpointVisitor.
- So for now we'll just get to a point where the CheckpointRemoverTask
  will be responsible only for (1) running periodically and (2)
  launching a VBCBAdaptor/CheckpointVisitor run.

In this patch I start with adding the logic that allows the
CheckpointVisitor to stop its execution based on how much memory it has
to release. For that, I add the CheckpointVisitor::memToRelease
member that will be updated at every VBucket visit. The Visitor will
stop when memToRelease=0.
The behaviour of the Visitor doesn't change yet in this patch as
memToReleased is initialized but never decremented.

Change-Id: I9670519740a6122dd877a617b2c518bbecd390c3
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/160152
Tested-by: Paolo Cocchi <paolo.cocchi@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>