MB-45654: Update purge seqno during Magma implicit compaction
Modify MagmaKVStore::compactionCallBack() to update a vbucket's purge
seqno if the we're performing a compaction in the implicit context. To
allow the method to be aware that we're in the implicit context set a
flag in the MagmaKVStore::MagmaCompactionCB after we've called
MagmaKVStore::makeCompactionContext(). Also add a lambda method to
CompactionContext so that we can bind on the EPBucket to give us access
to VBucket::setPurgeSeqno().

Change-Id: I230667da1bb9aff202e97ffc84cd51afc572043c
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/160292
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>