Reduce cost of including client_connection.h
client_connection.h is an expensive header to include, costing over 2s
(!) per inclusion:

    37896 ms: ../kv_engine/protocol/connection/client_connection.h (included 18 times, avg 2105 ms)

The main reason for the high cost is the various headers pulled in via
folly/AsyncSocket.h.

Reduce this by moving those headers to the .cc file, and using pimpl
and forward declarations of the needed types.

This reduces the overall build time by 61s (or 2.8%):

From:

    Compilation (677 times):
      Parsing (frontend):         1684.8 s
      Codegen & opts (backend):    520.0 s

To:

    Compilation (678 times):
      Parsing (frontend):         1643.5 s
      Codegen & opts (backend):    500.0 s

Change-Id: I113d0ad10c89be4c221f89362678bc8355e5e244
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/160322
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>