MB-47462: Introduce dedicated checkpoint destruction task
As a first step towards "eager" checkpoint removal, this patch
introduces a new CheckpointDestroyerTask.

This task receives ownership of checkpoints, and destroys them as soon
as possible. The task only runs when checkpoints have been queued for
destruction.

For now, this is a seemingly unnecessary level of indirection - the
CheckpointVisitor now queues checkpoints for a different task to
destroy, rather than destroying them immediately.

However, this separation will, in later patches, allow checkpoints to
be queued as soon as they become unreferenced (e.g., by cursor
movement). This will not require the CheckpointVisitor to scan all
vbuckets.

Change-Id: Iee7ba8a5a0b994f93f731584f548b04f195f6c0b
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/160717
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>