MB-47462: Shard checkpoint destruction task
Checkpoint destruction has previously been a "single threaded"
activity; a single task per bucket was responsible for it.

As load increases, with fast persistence/ephemeral buckets, destruction
of checkpoints may become a bottleneck, limiting the rate at which
memory will be recovered from CheckpointManagers.

To resolve this, add support for multiple CheckpointDestroyer tasks,
distributing checkpoints across them based on vbucket id.

This allows checkpoint removal to be coarsely scaled through the config
param `checkpoint_destruction_tasks`.

For now, the default is 1, matching the existing behaviour.

This may be increased in the future, but should be driven by
performance testing.

Change-Id: I3fc5511f61eabd12bcf3e810115e061471342444
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/160731
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>