MB-47460: Fix dropped/resurrected collection stats in compaction replay
Extending the tests created for MB-47460 revealed two bugs which are
addressed in this commit.

(1) A collection that is dropped and 'copied' to a datafile during
    compaction replay is leaving the stats document behind. It will be
    cleaned up in a subsequent compaction, but it's not consistent with
    a datafile which had a collection dropped and no compaction.

(2) A collection that is dropped and subsequently re-created during the
    replay combines the stats of the old and new generation of the
    collection.

For (1) replay commit now processes the dropped collections, removing
the stats documents of dropped collections.

For (2) FlushAccounting issues an erase of the collection from the
stats map whenever it encounters a collection event.

Change-Id: Ie5c1ecfbbd3d884f4813387de86d19f2ff129d0e
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/160912
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>